+++ butterfly/xlib/util/host_util.py
@@ -13,7 +13,7 @@ import socket
 import logging
 
 
-def find_host_of_ip(ip):
+def get_host_by_ip(ip):
     """找到机器名
     :param ip: IP, str
     :return host: str/None
@@ -28,7 +28,7 @@ def find_host_of_ip(ip):
         return None
 
 
-def find_ip_of_host(host):
+def get_ip_by_host(host):
     """找到机器的IP
     :param host: 机器名, str
     :return ip: str/None
@@ -58,32 +58,72 @@ def is_ip(host):
 if __name__ == '__main__':
     import sys
     import inspect
-    if len(sys.argv) < 2:
-        print "Usage:"
+
+    def _usage(func_name=""):
+        """
+        output the module usage
+        """
+        print("Usage:")
+        print("-------------------------------------------------")
         for k, v in sorted(globals().items(), key=lambda item: item[0]):
-            if inspect.isfunction(v) and k[0] != "_":
-                args, __, __, defaults = inspect.getargspec(v)
-                if defaults:
-                    print sys.argv[0], k, str(args[:-len(defaults)])[1:-1].replace(",", ""), \
-                        str(["%s=%s" % (a, b) for a, b in zip(args[-len(defaults):], defaults)])[1:-1].replace(",", "")
-                else:
-                    print sys.argv[0], k, str(v.func_code.co_varnames[:v.func_code.co_argcount])[1:-1].replace(",", "")
+            if func_name and func_name != k:
+                continue
+
+            if not inspect.isfunction(v) or k[0] == "_":
+                continue
+
+            args, __, __, defaults = inspect.getargspec(v)
+            #
+            # have defaults:
+            #       def hello(str_info, test="world"):
+            #               |
+            #               V
+            #       return: (args=['str_info', 'test'], varargs=None, keywords=None, defaults=('world',)
+            # no defaults:
+            #       def echo2(str_info1, str_info2):
+            #               |
+            #               V
+            #       return: (args=['str_info1', 'str_info2'], varargs=None, keywords=None, defaults=None)
+            #
+            # str(['str_info1', 'str_info2'])[1:-1].replace(",", "") ===> 'str_info1' 'str_info2'
+            #
+            if defaults:
+                args_all = str(args[:-len(defaults)])[1:-1].replace(",", ""), \
+                    str(["%s=%s" % (a, b) for a, b in zip(args[-len(defaults):], defaults)])[1:-1].replace(",", "")
+            else:
+                args_all = str(v.func_code.co_varnames[:v.func_code.co_argcount])[1:-1].replace(",", "")
+
+            if not isinstance(args_all, tuple):
+                args_all = tuple(args_all.split(" "))
+
+            exe_info = "{file_name} {func_name} {args_all}".format(
+                file_name=sys.argv[0],
+                func_name=k,
+                args_all=" ".join(args_all))
+            print(exe_info)
+
+            # output func_doc
+            if func_name and v.func_doc:
+                print("\n".join(["\t" + line.strip() for line in v.func_doc.strip().split("\n")]))
+
+        print("-------------------------------------------------")
+
+    if len(sys.argv) < 2:
+        _usage()
         sys.exit(-1)
     else:
         func = eval(sys.argv[1])
         args = sys.argv[2:]
         try:
             r = func(*args)
-            print r
-        except Exception as e:
-            print "Usage:"
-            print "\t", "python %s" % sys.argv[1], str(func.func_code.co_varnames[:func.func_code.co_argcount])[
-                1:-1].replace(",", "")
-            if func.func_doc:
-                print "\n".join(["\t\t" + line.strip() for line in func.func_doc.strip().split("\n")])
-            print e
+        except Exception:
+            _usage(func_name=sys.argv[1])
+
             r = -1
             import traceback
             traceback.print_exc()
+
         if isinstance(r, int):
             sys.exit(r)
+
+        print r